% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd.R
\name{psd_context}
\alias{psd_context}
\title{Return context: full source of this client}
\usage{
psd_context()
}
\value{
character: source code of this file
}
\description{
Return context: full source of this client
}
