% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd.R
\name{psd_list}
\alias{psd_list}
\title{List available Philadelphia School District datasets}
\usage{
psd_list()
}
\value{
tibble: id, name, url, category
}
\description{
List available Philadelphia School District datasets
}
