% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd.R
\name{psd_sheets}
\alias{psd_sheets}
\title{List sheets in the Excel files of a PSSA dataset}
\usage{
psd_sheets(dataset_id)
}
\arguments{
\item{dataset_id}{Dataset id from psd_list()}
}
\value{
named list: filename -> vector of sheet names
}
\description{
List sheets in the Excel files of a PSSA dataset
}
