% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd.R
\name{psd_zip_contents}
\alias{psd_zip_contents}
\title{List files inside a PSSA/Keystone ZIP archive}
\usage{
psd_zip_contents(dataset_id)
}
\arguments{
\item{dataset_id}{Dataset id from psd_list()}
}
\value{
tibble: Name, Length, Date (contents of ZIP)
}
\description{
List files inside a PSSA/Keystone ZIP archive
}
